Resize country list on login page, load registration countries dynamically
ISSUE: PLCR16-26
