+++ PacklinkPro/view/adminhtml/web/css/packlink.css
@@ -1177,6 +1177,11 @@
     width: 250px;
     float: right;
     animation: pl-fade-in .5s ease;
+    max-height: calc(100vh - 240px);
+    overflow-x: auto;
+    text-align: justify;
+    z-index: 1;
+    background-color: #fff;
 }
 
 .pl-register-country-title-wrapper {
@@ -1186,9 +1191,7 @@
 
 .pl-register-country-list-wrapper {
     margin-top: 15px;
-    max-height: 205px;
-    overflow-x: auto;
-    text-align: justify;
+    height: 97%;
 }
 
 .pl-register-country-list-wrapper a {
@@ -1197,14 +1200,15 @@
 
 .pl-country {
     display: flex;
-    height: 40px;
+    height: 35px;
     width: 30px;
-    line-height: 40px;
+    line-height: 35px;
     border-radius: 50%;
     text-align: center;
     color: #2095f2;
     font-size: 14px;
     text-transform: uppercase;
+    margin-bottom: 5px
 }
 
 .pl-country-logo {
